# frontend-react/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StockRL Research API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/research-previews": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Preview Research",
        "operationId": "preview_research_api_v1_research_previews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchProtocolDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/researches": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Submit Research",
        "operationId": "submit_research_api_v1_researches_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchProtocolDraft"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchSubmission"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Researches",
        "operationId": "researches_api_v1_researches_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/researches/{research_id}": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Research",
        "operationId": "research_api_v1_researches__research_id__get",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Research Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/researches/{research_id}/diagnostics": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Diagnostics",
        "operationId": "diagnostics_api_v1_researches__research_id__diagnostics_get",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Research Id"
            }
          },
          {
            "name": "instrument_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument Id"
            }
          },
          {
            "name": "fold_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fold Id"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 4294967295,
              "minimum": 0,
              "title": "Seed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticsReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/researches/{research_id}/resume": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Resume",
        "operationId": "resume_api_v1_researches__research_id__resume_post",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Research Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchResumeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AcceptanceThresholds": {
        "properties": {
          "minimum_folds": {
            "type": "integer",
            "const": 4,
            "title": "Minimum Folds",
            "default": 4
          },
          "required_seeds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Required Seeds",
            "default": [
              42,
              43,
              44,
              45,
              46
            ]
          },
          "comparable_fraction": {
            "type": "number",
            "const": 0.8,
            "title": "Comparable Fraction",
            "default": 0.8
          },
          "winning_fold_fraction": {
            "type": "number",
            "const": 0.6,
            "title": "Winning Fold Fraction",
            "default": 0.6
          },
          "maximum_drawdown_difference": {
            "type": "number",
            "const": 0.02,
            "title": "Maximum Drawdown Difference",
            "default": 0.02
          },
          "risk_absolute_tolerance": {
            "type": "number",
            "const": 0.02,
            "title": "Risk Absolute Tolerance",
            "default": 0.02
          },
          "risk_relative_tolerance": {
            "type": "number",
            "const": 0.2,
            "title": "Risk Relative Tolerance",
            "default": 0.2
          },
          "minimum_assets": {
            "type": "integer",
            "const": 5,
            "title": "Minimum Assets",
            "default": 5
          },
          "winning_asset_fraction": {
            "type": "number",
            "const": 0.6,
            "title": "Winning Asset Fraction",
            "default": 0.6
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AcceptanceThresholds"
      },
...
      "BudgetPreview": {
        "properties": {
          "unit_count": {
            "type": "integer",
            "title": "Unit Count"
          },
          "requested_total_steps": {
            "type": "integer",
            "title": "Requested Total Steps"
          },
          "rollout_upper_bound": {
            "type": "integer",
            "title": "Rollout Upper Bound"
          },
          "actual_steps_per_unit": {
            "type": "integer",
            "title": "Actual Steps Per Unit"
          },
          "rollout_steps": {
            "type": "integer",
            "title": "Rollout Steps"
          },
          "checkpoint_steps": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Checkpoint Steps"
          },
          "checkpoint_evaluations": {
            "type": "integer",
            "title": "Checkpoint Evaluations"
          },
          "calibration_evaluations": {
            "type": "integer",
            "title": "Calibration Evaluations"
          },
          "test_evaluations": {
            "type": "integer",
            "title": "Test Evaluations"
          },
          "extra_evaluations": {
            "type": "integer",
            "title": "Extra Evaluations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "unit_count",
          "requested_total_steps",
          "rollout_upper_bound",
          "actual_steps_per_unit",
          "rollout_steps",
          "checkpoint_steps",
          "checkpoint_evaluations",
          "calibration_evaluations",
          "test_evaluations",
          "extra_evaluations"
        ],
        "title": "BudgetPreview"
      },
...
      "DiagnosticsReport": {
        "properties": {
          "boundary_action_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary Action Ratio"
          },
          "clipping_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clipping Ratio"
          },
          "near_half_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Near Half Ratio"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          },
          "unavailable": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unavailable",
            "default": []
          },
          "feature_stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Feature Stats"
          },
          "regimes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Regimes"
          },
          "training_log": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Training Log",
            "default": []
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DiagnosticsReport"
      },
      "EmptyRequest": {
        "properties": {},
        "additionalProperties": false,
        "type": "object",
        "title": "EmptyRequest"
      },
...
      "FoldPlan": {
        "properties": {
          "fold_id": {
            "type": "string",
            "title": "Fold Id"
          },
          "instrument_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Id"
          },
          "train": {
            "$ref": "#/components/schemas/SessionWindow"
          },
          "validation": {
            "$ref": "#/components/schemas/SessionWindow"
          },
          "test": {
            "$ref": "#/components/schemas/SessionWindow"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fold_id",
          "train",
          "validation",
          "test"
        ],
        "title": "FoldPlan"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResearchAttemptResponse": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "title": "Attempt Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "attempt_id",
          "job_id",
          "status",
          "created_at"
        ],
        "title": "ResearchAttemptResponse"
      },
      "ResearchDetailResponse": {
        "properties": {
          "protocol_id": {
            "type": "string",
            "title": "Protocol Id"
          },
          "technical_status": {
            "type": "string",
            "enum": [
              "completed",
              "failed",
              "cancelled",
              "incomplete"
            ],
            "title": "Technical Status"
          },
          "qualification": {
            "type": "string",
            "enum": [
              "exploratory",
              "declared_holdout"
            ],
            "title": "Qualification"
          },
          "economic_outcome": {
            "type": "string",
            "enum": [
              "insufficient_evidence",
              "no_added_value",
              "candidate_edge"
            ],
            "title": "Economic Outcome"
          },
          "provisional_outcome": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "insufficient_evidence",
                  "no_added_value",
                  "candidate_edge"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Provisional Outcome"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons",
            "default": []
          },
          "assets": {
            "additionalProperties": true,
            "type": "object",
            "title": "Assets"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/UnitResult"
            },
            "type": "array",
            "title": "Units",
            "default": []
          },
          "research_id": {
            "type": "string",
            "title": "Research Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "job_status": {
            "$ref": "#/components/schemas/JobDetail"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "protocol": {
            "$ref": "#/components/schemas/ResearchProtocol"
          },
          "protocol_sha256": {
            "type": "string",
            "title": "Protocol Sha256"
          },
          "preview": {
            "$ref": "#/components/schemas/ResearchPreviewResponse"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/ResearchAttemptResponse"
            },
            "type": "array",
            "title": "Attempts"
          },
          "unit_progress": {
            "items": {
              "$ref": "#/components/schemas/ResearchUnitProgress"
            },
            "type": "array",
            "title": "Unit Progress"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "protocol_id",
          "technical_status",
          "qualification",
          "economic_outcome",
          "research_id",
          "job_id",
          "job_status",
          "created_at",
          "protocol",
          "protocol_sha256",
          "preview",
          "attempts",
          "unit_progress"
        ],
        "title": "ResearchDetailResponse"
      },
      "ResearchListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResearchSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "has_more",
          "next_cursor"
        ],
        "title": "ResearchListResponse"
      },
      "ResearchPreviewResponse": {
        "properties": {
          "fold_plan": {
            "items": {
              "$ref": "#/components/schemas/FoldPlan"
            },
            "type": "array",
            "title": "Fold Plan"
          },
          "budget": {
            "$ref": "#/components/schemas/BudgetPreview"
          },
          "checkpoint_steps": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Checkpoint Steps"
          },
          "qualification": {
            "type": "string",
            "enum": [
              "exploratory",
              "declared_holdout"
            ],
            "title": "Qualification"
          },
          "blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blockers",
            "default": []
          },
          "omitted_folds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Omitted Folds",
            "default": []
          },
          "exposure_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exposure Warnings"
          },
          "protocol_sha256": {
            "type": "string",
            "title": "Protocol Sha256"
          },
          "canonical_request": {
            "$ref": "#/components/schemas/ResearchProtocolDraft"
          },
          "maximum_runtime_seconds": {
            "type": "integer",
            "title": "Maximum Runtime Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fold_plan",
          "budget",
          "checkpoint_steps",
          "qualification",
          "exposure_warnings",
          "protocol_sha256",
          "canonical_request",
          "maximum_runtime_seconds"
        ],
        "title": "ResearchPreviewResponse"
      },
...
      "ResearchProtocolDraft": {
        "properties": {
          "protocol_id": {
            "type": "string",
            "minLength": 1,
            "title": "Protocol Id"
          },
          "version": {
            "type": "integer",
            "const": 2,
            "title": "Version",
            "default": 2
          },
          "hypothesis": {
            "type": "string",
            "title": "Hypothesis",
            "default": "Can autonomous RL allocation improve net performance consistently against transparent references at comparable risk?"
          },
          "scope": {
            "type": "string",
            "enum": [
              "single_asset",
              "asset_set"
            ],
            "title": "Scope",
            "default": "single_asset"
          },
          "instrument_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Instrument Ids"
          },
          "dataset_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dataset Ids"
          },
          "dataset_fingerprints": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Dataset Fingerprints"
          },
          "market_profile_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Market Profile Ids"
          },
          "market_profile_fingerprints": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Market Profile Fingerprints"
          },
          "asset_selection_note": {
            "type": "string",
            "minLength": 1,
            "title": "Asset Selection Note"
          },
          "first_test_session": {
            "type": "string",
            "title": "First Test Session"
          },
          "fold_count": {
            "type": "integer",
            "maximum": 250.0,
            "exclusiveMinimum": 0.0,
            "title": "Fold Count",
            "default": 4
          },
          "seeds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Seeds",
            "default": [
              42,
              43,
              44,
              45,
              46
            ]
          },
          "training_budget": {
            "$ref": "#/components/schemas/TrainingBudget"
          },
          "checkpoint_budget": {
            "type": "integer",
            "maximum": 20.0,
            "exclusiveMinimum": 0.0,
            "title": "Checkpoint Budget",
            "default": 20
          },
          "primary_reference_ids": {
            "items": {
              "type": "string",
              "enum": [
                "fixed_50",
                "matched_fixed"
              ]
            },
            "type": "array",
            "title": "Primary Reference Ids",
            "default": [
              "fixed_50",
              "matched_fixed"
            ]
          },
          "cost_scenarios": {
            "items": {
              "type": "string",
              "enum": [
                "base",
                "execution_x2",
                "execution_x3"
              ]
            },
            "type": "array",
            "title": "Cost Scenarios",
            "default": [
              "base",
              "execution_x2",
              "execution_x3"
            ]
          },
          "qualification": {
            "type": "string",
            "enum": [
              "exploratory",
              "declared_holdout"
            ],
            "title": "Qualification",
            "default": "exploratory"
          },
          "acceptance_thresholds": {
            "$ref": "#/components/schemas/AcceptanceThresholds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "protocol_id",
          "instrument_ids",
          "dataset_ids",
          "dataset_fingerprints",
          "market_profile_ids",
          "market_profile_fingerprints",
          "asset_selection_note",
          "first_test_session"
        ],
        "title": "ResearchProtocolDraft"
      },
      "ResearchResumeResponse": {
        "properties": {
          "research_id": {
            "type": "string",
            "title": "Research Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "research_id",
          "job_id"
        ],
        "title": "ResearchResumeResponse"
      },
      "ResearchSubmission": {
        "properties": {
          "research_id": {
            "type": "string",
            "title": "Research Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "protocol_sha256": {
            "type": "string",
            "title": "Protocol Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "research_id",
          "job_id",
          "protocol_sha256"
        ],
        "title": "ResearchSubmission"
      },
      "ResearchSummary": {
        "properties": {
          "research_id": {
            "type": "string",
            "title": "Research Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "technical_status": {
            "type": "string",
            "title": "Technical Status"
          },
          "protocol_sha256": {
            "type": "string",
            "title": "Protocol Sha256"
          },
          "qualification": {
            "type": "string",
            "enum": [
              "exploratory",
              "declared_holdout"
            ],
            "title": "Qualification"
          },
          "hypothesis": {
            "type": "string",
            "title": "Hypothesis"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "research_id",
          "job_id",
          "created_at",
          "technical_status",
          "protocol_sha256",
          "qualification",
          "hypothesis"
        ],
        "title": "ResearchSummary"
      },
...
      "TrainingBudget": {
        "properties": {
          "requested_timesteps": {
            "type": "integer",
            "maximum": 1000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Requested Timesteps",
            "default": 100000
          },
          "episode_length": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Episode Length",
            "default": 126
          },
          "algorithm": {
            "type": "string",
            "enum": [
              "PPO",
              "SAC"
            ],
            "title": "Algorithm",
            "default": "PPO"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TrainingBudget"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}